--- src/main/kotlin/lotto/model/RandomNumber.kt
@@ -1,7 +1,0 @@
-package lotto.model
-
-object RandomNumber {
-    fun getLottoNum(): Int {
-        return (1..45).random()
-    }
-}
